Extract MockRails to share across other test files
setup/teardown is still manual for now
